Extract calorie functions Extracts all calorie frunctions from diary.js to calories.js. Modifies all function invocations in diary.js to appropriately call module exports from calories.js. All page functionality remains as it was prior to refactor.
