+++ lib/diary/diary-food-table.js
@@ -16,7 +16,6 @@ function  getFoods() {
   })
 }
 
-
 function makeFoodTable(data) {
   data.forEach( (food) => {
     $('#diary-foods-body').prepend(`<tr data-food-id=${food.id}><td class='food-cell' >${food.name}</td><td class='calorie-cell'>${food.calories}</td><td><input type="checkbox" id="${food.id}" data-food-name="${food.name}"/><label for="${food.id}"></label></td></tr>`)
