+++ lib/diary/diary.js
@@ -1,9 +1,9 @@
 const API = 'https://qs-be.herokuapp.com/api/v1/'
 const LOCAL = 'http://localhost:3000/api/v1/'
-const DiaryFoodTable = require('./diary-food-table')
+const diaryFoodTable = require('./diary-food-table')
+const calories = require('./calories')
 $ = require('jquery')
 
-const mealNames = ["Breakfast", "Lunch", "Dinner", "Snacks"]
 let counter = 0
 
 
@@ -56,7 +56,6 @@ const Diary = class Diary {
         console.error(error)
       });
     })
-
   }
 
   getPrevious() {
@@ -78,7 +77,7 @@ const Diary = class Diary {
     $('#Snacks-body').empty()
     $('#Breakfast-body').empty()
 
-    resetCalories()
+    calories.resetCalories()
 
     $('#date-container').empty()
     $('#date-container').append(date)
@@ -115,7 +114,7 @@ const Diary = class Diary {
     $('#Snacks-body').empty()
     $('#Breakfast-body').empty()
 
-    resetCalories()
+    calories.resetCalories()
 
     $('#date-container').empty()
     $('#date-container').append(date)
@@ -134,70 +133,20 @@ const Diary = class Diary {
   }
 }
 
-function resetCalories() {
-  $('#Breakfast-total-calories').text(0)
-  $('#Breakfast-remaining-calories').css('color', 'green').text(400)
-
-  $('#Dinner-total-calories').text(0)
-  $('#Dinner-remaining-calories').css('color', 'green').text(800)
-
-  $('#Lunch-total-calories').text(0)
-  $('#Lunch-remaining-calories').css('color', 'green').text(600)
-
-  $('#Snacks-total-calories').text(0)
-  $('#Snacks-remaining-calories').css('color', 'green').text(200)
-
-  $('#consumed-calories').text(0)
-  $('#remaining-calories').css('color', 'green').text(2000)
-}
-
-function updateMealRemainingCalories(mealName, food){
-  let currentValue = parseInt($(`#${mealName}-remaining-calories`).text())
-  let updatedValue = currentValue - food.calories
-  $(`#${mealName}-remaining-calories`).text(updatedValue)
-  if(updatedValue < 0) {
-    $(`#${mealName}-remaining-calories`).css('color', 'red')
-  } else if (updatedValue >= 0) {
-    $(`#${mealName}-remaining-calories`).css('color', 'green')
-  }
-}
-
 function addFoodstoMealTable(meal, mealName){
 
   if (Array.isArray(meal.foods)) {
     meal.foods.forEach( (food) => {
       $(`#${mealName}-body`).prepend(`<tr><td >${food.name}</td><td id='${mealName}-calories'>${food.calories}</td><td><button data-meal-id=${food.meal_id} class='remove btn red darken-3'>-</button></td></tr>`)
-      updateMealTotalCalories(mealName, food)
-      updateMealRemainingCalories(mealName, food)
+      calories.updateMealTotalCalories(mealName, food)
+      calories.updateMealRemainingCalories(mealName, food)
     })
-    updateConsumedCalories()
+    calories.updateConsumedCalories()
   }
 }
 
-function updateConsumedCalories(){
-  $('#consumed-calories').text(0)
-    mealNames.forEach( (mealName) => {
-      let mealCalories = parseInt($(`#${mealName}-total-calories`).text())
-      let totalCalories = parseInt($('#consumed-calories').text())
-      $('#consumed-calories').text(`${mealCalories + totalCalories}`)
-    })
-    updateRemainingCalories()
-}
-
-function updateRemainingCalories(){
-  let consumedCalories = parseInt($('#consumed-calories').text())
-  $('#remaining-calories').text(`${2000 - consumedCalories}`)
-  if ((2000 - consumedCalories) < 0) {
-    $('#remaining-calories').css('color', 'red')
-  }
-  else if (2000 - consumedCalories >= 0) {
-    $('#remaining-calories').css('color', 'green')
-
-  }
-
-}
 function populateMealsTables(data){
-  resetCalories()
+  calories.resetCalories()
   if (Array.isArray(data)) {
     $('#date-box').attr('data-diary-id', `${data[0].diary_id}`)
     data.forEach((meal) => {
@@ -207,12 +156,6 @@ function populateMealsTables(data){
   } else {
     $('#date-box').attr('data-diary-id', `${data.id}`)
   }
-}
-
-function updateMealTotalCalories(mealName, food){
-  let currentValue = parseInt($(`#${mealName}-total-calories`).text())
-  let updatedValue = currentValue + food.calories
-  $(`#${mealName}-total-calories`).text(updatedValue)
 }
 
 function grabFood(data) {
@@ -234,10 +177,10 @@ function addFoodToMeal(data, mealName, meal_id){
   const label = $(`#${data.id}`).next()
   $(label).prop('checked', false)
   $(`#${mealName}-body`).prepend(`<tr><td >${data.name}</td><td id='${mealName}-calories'>${data.calories}</td><td><button data-meal-id=${meal_id} class='remove btn red darken-3'>-</button></td></tr>`)
-  updateMealTotalCalories(mealName, data)
-  updateMealRemainingCalories(mealName, data)
-  updateRemainingCalories()
-  updateConsumedCalories()
+  calories.updateMealTotalCalories(mealName, data)
+  calories.updateMealRemainingCalories(mealName, data)
+  calories.updateRemainingCalories()
+  calories.updateConsumedCalories()
 }
 
 function searchFoods(){
@@ -255,18 +198,16 @@ function searchFoods(){
   })
 }
 
+function removeFoodFromMealTable(button) {
+  let mealName = $(button).offsetParent().find('span').text()
+  let calorieUpdate = -parseInt($(button).parent().prev().text())
+  $(button).parent().parent().remove()
 
-
-  function removeFoodFromMealTable(button) {
-    let mealName = $(button).offsetParent().find('span').text()
-    let calories = -parseInt($(button).parent().prev().text())
-    $(button).parent().parent().remove()
-
-    updateMealTotalCalories(mealName, {calories: calories})
-    updateMealRemainingCalories(mealName, {calories: calories})
-    updateRemainingCalories()
-    updateConsumedCalories()
-  }
+  calories.updateMealTotalCalories(mealName, {calories: calorieUpdate})
+  calories.updateMealRemainingCalories(mealName, {calories: calorieUpdate})
+  calories.updateRemainingCalories()
+  calories.updateConsumedCalories()
+}
 
 function removeFoodFromMeal(){
   const button = event.target
@@ -288,7 +229,7 @@ $(document).ready(function(){
     searchFoods()
   })
 
-  DiaryFoodTable.getFoods()
+  diaryFoodTable.getFoods()
 
   diary.getDiary()
